Move folder and change to ScopeAwareRector
