+++ rules-tests/Php83/Rector/ClassMethod/AddOverrideAttributeToOverriddenMethodsRector/Fixture/apply_attribute_to_overridden_methods_from_grandparent.php.inc
@@ -1,0 +1,30 @@
+<?php
+
+namespace Rector\Tests\Php83\Rector\ClassMethod\AddOverrideAttributeToOverriddenMethodsRector\Fixture;
+
+use Rector\Tests\Php83\Rector\ClassMethod\AddOverrideAttributeToOverriddenMethodsRector\Source\ExampleChildClass;
+
+class ChildClass extends ExampleChildClass
+{
+    public function foo()
+    {
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\Php83\Rector\ClassMethod\AddOverrideAttributeToOverriddenMethodsRector\Fixture;
+
+use Rector\Tests\Php83\Rector\ClassMethod\AddOverrideAttributeToOverriddenMethodsRector\Source\ExampleChildClass;
+
+class ChildClass extends ExampleChildClass
+{
+    #[\Override]
+    public function foo()
+    {
+    }
+}
+
+?>
