+++ rules-tests/Php83/Rector/ClassMethod/AddOverrideAttributeToOverriddenMethodsRector/Fixture/skip_override_attribute_exists.php.inc
@@ -1,0 +1,15 @@
+<?php
+
+namespace Rector\Tests\Php83\Rector\ClassMethod\AddOverrideAttributeToOverriddenMethodsRector\Fixture;
+
+use Rector\Tests\Php83\Rector\ClassMethod\AddOverrideAttributeToOverriddenMethodsRector\Source\ExampleParentClass;
+
+class ChildClass extends ExampleParentClass
+{
+    #[\Override]
+    public function foo()
+    {
+    }
+}
+
+?>
