+++ rules-tests/Php83/Rector/ClassMethod/AddOverrideAttributeToOverriddenMethodsRector/Source/ExampleChildClass.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rector\Tests\Php83\Rector\ClassMethod\AddOverrideAttributeToOverriddenMethodsRector\Source;
+
+class ExampleChildClass extends ExampleParentClass
+{
+}
