+++ rules-tests/Php83/Rector/ClassMethod/AddOverrideAttributeToOverriddenMethodsRector/Source/ExampleConstructParentClass.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rector\Tests\Php83\Rector\ClassMethod\AddOverrideAttributeToOverriddenMethodsRector\Source;
+
+class ExampleConstructParentClass
+{
+    public function __construct()
+    {
+    }
+}
