+++ rules-tests/Php83/Rector/ClassMethod/AddOverrideAttributeToOverriddenMethodsRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\Php83\Rector\Class_\AddOverrideAttributeToOverriddenMethodsRector;
+use Rector\Php83\Rector\ClassMethod\AddOverrideAttributeToOverriddenMethodsRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(AddOverrideAttributeToOverriddenMethodsRector::class);
