--- rules-tests/Php83/Rector/Class_/AddOverrideAttributeToOverriddenMethodsRector/Fixture/skip_method_that_is_not_overridden.php.inc
@@ -1,14 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php83\Rector\Class_\AddOverrideAttributeToOverriddenMethodsRector\Fixture;
-
-use Rector\Tests\Php83\Rector\Class_\AddOverrideAttributeToOverriddenMethodsRector\Source\ExampleParentClass;
-
-class ChildClass extends ExampleParentClass
-{
-    public function bar()
-    {
-    }
-}
-
-?>
