--- rules-tests/Php83/Rector/Class_/AddOverrideAttributeToOverriddenMethodsRector/Fixture/skip_override_attribute_exists.php.inc
@@ -1,15 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php83\Rector\Class_\AddOverrideAttributeToOverriddenMethodsRector\Fixture;
-
-use Rector\Tests\Php83\Rector\Class_\AddOverrideAttributeToOverriddenMethodsRector\Source\ExampleParentClass;
-
-class ChildClass extends ExampleParentClass
-{
-    #[\Override]
-    public function foo()
-    {
-    }
-}
-
-?>
