--- rules-tests/Php83/Rector/Class_/AddOverrideAttributeToOverriddenMethodsRector/Source/ExampleChildClass.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php83\Rector\Class_\AddOverrideAttributeToOverriddenMethodsRector\Source;
-
-class ExampleChildClass extends ExampleParentClass
-{
-}
