--- rules-tests/Php83/Rector/Class_/AddOverrideAttributeToOverriddenMethodsRector/Source/ExampleConstructParentClass.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php83\Rector\Class_\AddOverrideAttributeToOverriddenMethodsRector\Source;
-
-class ExampleConstructParentClass
-{
-    public function __construct()
-    {
-    }
-}
